Optimize rwlock uncontended rdlock
CAS upfront with the expectation that the lock isn't contended.
This will either pass or fail, fail will give us its currently value.
We can then use that current value to try a contended rdlock with other
readers.
Failing that it falls to the loop implementation.

This is faster since we need to already do an atomic load, might as well
as be a CAS.